requirements.txt
aiogram
tortoise-orm
aiosqlite
fastapi
uvicorn
jinja2
python-multipart
